+++ internal/jsonx/doc.go
@@ -1,0 +1,3 @@
+// Package jsonx contains utilities for dealing with JSON and the encoding/json
+// package.
+package jsonx